tourism-dwh/omd_requirements.txt
openmetadata-ingestion==1.10.10.0
